% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lmer_estimates.R
\name{calculate_lmer_estimates}
\alias{calculate_lmer_estimates}
\title{calculate_lmer_estimates}
\usage{
calculate_lmer_estimates(
  data,
  condition_column,
  experimental_columns,
  response_column,
  target_columns,
  condition_is_categorical,
  response_is_categorical = FALSE,
  output = NULL
)
}
\arguments{
\item{data}{Input data}

\item{condition_column}{Name of the condition variable (ex variable with values such as control/case). The input file has to have a corresponding column name}

\item{experimental_columns}{Name of variables related to experimental design such as "experiment", "plate", and "cell_line".}

\item{response_column}{Name of the variable observed by performing the experiment. ex) intensity.}

\item{target_columns}{Name of the experimental parameters to use for the power calculation.}

\item{condition_is_categorical}{Specify whether the condition variable is categorical. TRUE: Categorical, FALSE: Continuous.}

\item{response_is_categorical}{Default: the observed variable is continuous Categorical response variable will be implemented in the future. TRUE: Categorical , FALSE: Continuous (default).}

\item{output}{Output file name}
}
\value{
A linear mixed model result
}
\description{
This function performs a linear mixed model analysis using lmer.

Note: The current version does not accept categorical response variables, sample size parameters smaller than the observed samples size
}
