% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_power.R
\name{calculate_power}
\alias{calculate_power}
\title{calculate_power}
\usage{
calculate_power(
  data,
  power_curve = 0,
  variance_estimate_from = "data",
  condition_variable,
  experimental_variable,
  response_variable,
  categorical_condition_variable = 1,
  categorical_response_variable = 0,
  nsimn = 1000,
  target_parameters,
  levels = NULL,
  max_size = NULL,
  breaks = NULL,
  effect_size = NULL,
  ICC = NULL,
  output = NULL
)
}
\arguments{
\item{data}{Input data}

\item{power_curve}{1: Power simulation over a range of sample sizes or levels. 0: Power calculation over a single sample size or a level (default).}

\item{variance_estimate_from}{"data" (defalut) or "ICC". "data": If variances should be calculated from the input data. "ICC": User should provide Intra-Class Coefficients (ICC) in addition to observed data}

\item{condition_variable}{Name of the condition varaible (ex variable with values such as control/case). The input file has to have a corresponding column name}

\item{experimental_variable}{Name of the varaibles related to experimental desgin such as "experiment", "plate", and "cell_line".}

\item{response_variable}{Name of the variable observed by performing the experiment. ex) intensity.}

\item{categorical_condition_variable}{Specify whether the condition variable is categorical. 1: Categorical (default), 0: Continuous.}

\item{categorical_response_variable}{Default: the observed variable is continous. Categorical response variable will be implemented in the future. 1: Categorical , 0: Continuous (default).}

\item{levels}{1: Amplify the number of corresponding target parameter. 0: Amplify the number of samples from the corresponding target parameter, ex) If target_parameter = c("experiment","cell_line") and if you want to expand the number of experiment and sample more cells from each cell line, levels = c(1,0).}

\item{max_size}{Maximum levels or sample sizes to test. Default: the current level or the current sample size x 5. ex) If max_levels = c(10,5), it will test upto 10 experiments and 5 cell lines.}

\item{breaks}{Levels /sample sizes of the variable to be specified along the power curve.. Default: max(1, round( the number of current levels / 5 ))}

\item{effect_size}{If you know the effect size of your condition variable, provided it. If the effect size is not provided, it will be estimated from your data}

\item{ICC}{Intra-Class Coefficients (ICC) for each parameter}

\item{output}{Output file name}

\item{nsim}{The number of simulations to run. Default=1000}

\item{target_parameter}{Name of the experimental parameter to use for the power calculation.}
}
\value{
A power curve image or a power calculation result printed in a text file
}
\description{
This function uses simulation to perform power analysis. It is designed to explore the power of biological experiments and to suggest an optimal number of experimental variables

with reasonable power. The backbone of the function is based on simr package, which fits a fixed effect or mixed effect model based on the observed data and simulates response

variables. Users can test the power of different combinations of experimental variables and parameters.

Note: The current version does not accept categorical response variables, sample size parameters smaller than the observed samples size
}
\examples{
calculate_power(power_curve=1,data=inputData,
variance_estimate_from="data",condition_variable="classification",
experimental_variable=c("experiment","plate","line"), response_variable="edge.detection",
nsimn=10, target_parameters="experiment", levels=1)
}
