% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{check_normality}
\alias{check_normality}
\title{check_normality}
\usage{
check_normality(
  data,
  condition_column,
  experimental_columns,
  response_column,
  condition_is_categorical,
  response_is_categorical = FALSE,
  image_title = ""
)
}
\description{
check_normality
}
