% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_residuals_covariate.R
\name{get_residuals_covariate}
\alias{get_residuals_covariate}
\title{get_residuals_covariate}
\usage{
get_residuals_covariate(
  data,
  condition_column,
  experimental_columns,
  response_column,
  condition_is_categorical,
  covariate = NA,
  repeatable_columns = NA,
  response_is_categorical = FALSE,
  family_p = NULL,
  na.action = "complete"
)
}
\arguments{
\item{data}{Input data}

\item{condition_column}{Name of the condition variable (ex variable with values such as control/case). The input file has to have a corresponding column name}

\item{experimental_columns}{Name of variables related to experimental design such as "experiment", "plate", and "cell_line". They should be in order, for example, "experiment" should always come first .}

\item{response_column}{Name of the variable observed by performing the experiment. ex) intensity.}

\item{condition_is_categorical}{Specify whether the condition variable is categorical. TRUE: Categorical, FALSE: Continuous.}

\item{covariate}{The name of the covariate to control in the regression model}

\item{repeatable_columns}{Name of experimental variables that may appear repeatedly with the same ID. For example, cell_line C1 may appear in multiple experiments, but plate P1 cannot appear in more than one experiment}

\item{response_is_categorical}{Default: the observed variable is continuous Categorical response variable will be implemented in the future. TRUE: Categorical , FALSE: Continuous (default).}

\item{family_p}{The type of distribution family to specify when the response is categorical. If family is "binary" then binary(link="log") is used, if family is "poisson" then poisson(link="logit") is used, if family is "poisson_log" then poisson(link=") log") is used.}

\item{na.action}{"complete": missing data is not allowed in all columns (default), "unique": missing data is not allowed only in condition, experimental, and response columns. Selecting "complete" removes an entire row when there is one or more missing values, which may affect the distribution of other features.}
}
\value{
The adjusted residual values for the experimental variables will be returned along with the original input data. A boxplot of residual values and a plot of 95% confidence interval mean residual values adjusted for the experimental variables will be returned.
}
\description{
This function retrieve residual values from an lmerFit summary object and plot residual values by condition_column

Note: The current version does not accept categorical response variables, sample size parameters smaller than the observed samples size
}
\examples{
result=get_residuals_covariate(data=RMeDPower_data1,
condition_column="classification",
experimental_columns=c("experiment", "line"),
response_column="cell_size1",
condition_is_categorical=TRUE,
covariate="covariate",
repeatable_columns = "line",
response_is_categorical=FALSE)
}
