% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_qqplot_function.R, R/paste_columns.R
\name{paste_columns}
\alias{paste_columns}
\title{outlier_qqplot_function}
\usage{
paste_columns(data1, data2)

paste_columns(data1, data2)
}
\description{
outlier_qqplot_function

paste_columns
}
