% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_experimental_variables.R
\name{rename_experimental_variables}
\alias{rename_experimental_variables}
\title{rename_experimental_variables}
\usage{
rename_experimental_variables(data, experimental_columns)
}
\description{
rename_experimental_variables
}
