% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{transform_data}
\usage{
transform_data(
  data,
  condition_column,
  experimental_columns,
  response_column,
  condition_is_categorical,
  response_is_categorical = FALSE
)
}
\description{
transform_data
}
