% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data2_covariate.R
\name{transform_data2_covariate}
\alias{transform_data2_covariate}
\title{transform_data2_covariate}
\usage{
transform_data2_covariate(
  data,
  condition_column,
  experimental_columns,
  response_column,
  condition_is_categorical,
  covariate = NA,
  method = "rosner",
  repeatable_columns = NA,
  response_is_categorical = FALSE,
  family_p = NULL,
  alpha = 0.05,
  na.action = "complete"
)
}
\arguments{
\item{data}{Input data}

\item{condition_column}{Name of the condition variable (ex variable with values such as control/case). The input file has to have a corresponding column name}

\item{experimental_columns}{Name of the variable related to experimental design such as "experiment", "plate", and "cell_line". They should be in order, for example, "experiment" should always come first .}

\item{response_column}{Name of the variable observed by performing the experiment. ex) intensity.}

\item{condition_is_categorical}{Specify whether the condition variable is categorical. TRUE: Categorical, FALSE: Continuous.}

\item{covariate}{The name of the covariate to control in the regression model}

\item{method}{The method used to detect outliers. "rosner" (default) runs Rosner's test and "cook" runs Cook's distance.}

\item{repeatable_columns}{Name of experimental variables that may appear repeatedly with the same ID. For example, cell_line C1 may appear in multiple experiments, but plate P1 cannot appear in more than one experiment}

\item{response_is_categorical}{Default: the observed variable is continuous Categorical response variable will be implemented in the future. TRUE: Categorical , FALSE: Continuous (default).}

\item{family_p}{The type of distribution family to specify when the response is categorical. If family is "binary" then binary(link="log") is used, if family is "poisson" then poisson(link="logit") is used, if family is "poisson_log" then poisson(link=") log") is used.}

\item{alpha}{numeric scalar between 0 and 1 indicating the Type I error associated with the test of outliers}

\item{na.action}{"complete": missing data is not allowed in all columns (default), "unique": missing data is not allowed only in condition, experimental, and response columns. Selecting "complete" removes an entire row when there is one or more missing values, which may affect the distribution of other features.}
}
\value{
For continuous data, the function returns quantile-quanitle (qq) plots of i) raw residual values ii) log-transformed residual values iii) raw residual values after removing outliers, and iv) log-transformed residual values. For discrete data, it returns a histogram. If "rosner" is chosen, a matrix with updated feature values after transformation will be returned. If "cook" is choose, a list with  a matrix with updated feature values after transformation will be returned,
}
\description{
This functions makes quantile-quanitle (qq) plots of i) raw residual values ii) log-transformed residual values iii) raw residual values after removing outliers, and iv) log-transformed residual values after removing outliers. To detect outliers, the function uses Rosner's test.
}
\examples{
result=transform_data2(data=data, condition_column="classif", experimental_columns=c("experiment","line"), response_column="feature", condition_is_categorical=TRUE, response_is_categorical=FALSE, alpha=0.05, repeatable_columns = "line", method="cook", na.action="complete")
result=transform_data2(data=data, condition_column="classif", experimental_columns=c("experiment","line"), response_column="feature", condition_is_categorical=TRUE, response_is_categorical=FALSE, alpha=0.05, repeatable_columns = "line", method="cook", na.action="complete")
result=transform_data2(data=data, condition_column="classif", experimental_columns=c("experiment","line"), response_column="feature", condition_is_categorical=TRUE, response_is_categorical=TRUE, family_p="poisson", alpha=0.05, repeatable_columns = "line", method="cook", na.action="complete")
}
