% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data_by_residual.R
\name{transform_data_by_residual}
\alias{transform_data_by_residual}
\title{transform_data_by_residual}
\usage{
transform_data_by_residual(
  data,
  residual_values,
  response_column,
  alpha = 0.05
)
}
\arguments{
\item{data}{Input data}

\item{residual_values}{User provided residual values}

\item{response_column}{Name of the variable observed by performing the experiment. ex) intensity.}

\item{alpha}{numeric scalar between 0 and 1 indicating the Type I error associated with the test of outliers}
}
\value{
Quantile-quanitle (qq) plots of i) raw residual values ii) log-transformed residual values iii) raw residual values after removing outliers, and iv) log-transformed residual values
}
\description{
This functions makes quantile-quanitle (qq) plots of i) raw residual values ii) log-transformed residual values iii) raw residual values after removing outliers, and iv) log-transformed residual values after removing outliers. To detect outliers, the function uses Rosner's test.
}
\examples{
transform_data(data,"classif",c("experiment","line"),"feature1","TRUE")
}
