% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data_covariate.R
\name{transform_data_covariate}
\alias{transform_data_covariate}
\title{transform_data_covariate}
\usage{
transform_data_covariate(
  data,
  condition_column,
  experimental_columns,
  response_column,
  condition_is_categorical,
  covariate = NA,
  repeatable_columns = NA,
  response_is_categorical = FALSE,
  alpha = 0.05,
  na.action = "complete"
)
}
\arguments{
\item{data}{Input data}

\item{condition_column}{The name of the condition variable (ex a variable with values such as control/case). The input file has to have a corresponding column name}

\item{experimental_columns}{Names of variables related to the experimental design, such as "experiment", "plate", and "cell_line". They should be in order, for example, "experiment" should always come first .}

\item{response_column}{The name of the variable observed by performing the experiment. ex) intensity.}

\item{condition_is_categorical}{Specify whether the condition variable is categorical. TRUE: Categorical, FALSE: Continuous.}

\item{covariate}{The name of the covariate to control in the regression model}

\item{repeatable_columns}{Name of experimental variables that may appear repeatedly with the same ID. For example, cell_line C1 may appear in multiple experiments, but plate P1 cannot appear in more than one experiment}

\item{response_is_categorical}{Default: Observed variable is continuous. Categorical response variable will be implemented in the future. TRUE: Categorical , FALSE: Continuous (default).}

\item{alpha}{A numeric scalar between 0 and 1 representing the type I error associated with the outlier test.}

\item{na.action}{"complete": missing data is not allowed in all columns (default), "unique": missing data is not allowed only in condition, experimental, and response columns. Selecting "complete" removes an entire row when there is one or more missing values, which may affect the distribution of other features.}
}
\value{
Quantile-quanitle (qq) plots of i) raw residual values ii) log-transformed residual values iii) raw residual values after removing outliers, and iv) log-transformed residual values

This function returns a matrix with original columns and additional columns with transformed values. Users can select a feature column for power analysis or regression analysis based on QQ plot results.
}
\description{
This function makes quantile-quanitle (qq) plots of i) raw residual values ii) log-transformed residual values iii) raw residual values after removing outliers, and iv) log-transformed residual values after removing outliers. To detect outliers, the function uses Rosner's test.
}
\examples{
result=transform_data(data=RMeDPower_data1,
condition_column="classification",
experimental_columns=c("experiment", "line"),
response_column="cell_size1",
condition_is_categorical=TRUE,
repeatable_columns = "line",
response_is_categorical=FALSE)
}
