% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{check_normality}
\alias{check_normality}
\title{check_normality}
\usage{
check_normality(
  data,
  condition_column,
  experimental_columns,
  response_column,
  condition_is_categorical,
  response_is_categorical = FALSE
)
}
\arguments{
\item{data}{Input data}

\item{condition_column}{Name of the condition variable (ex variable with values such as control/case). The input file has to have a corresponding column name}

\item{experimental_columns}{Name of the variable related to experimental design such as "experiment", "plate", and "cell_line".}

\item{response_column}{Name of the variable observed by performing the experiment. ex) intensity.}

\item{condition_is_categorical}{Specify whether the condition variable is categorical. TRUE: Categorical, FALSE: Continuous.}

\item{response_is_categorical}{Default: the observed variable is continuous Categorical response variable will be implemented in the future. TRUE: Categorical , FALSE: Continuous (default).}
}
\value{
A quantile-quantile (qq) plot of residual values in a mixed-effects model and optional figure saved in png
}
\description{
This function makes a quantile-quantile (qq) plot of the residual values of the mixed effects model.

Users can check the normality of residual values by examining qqplot.
}
\examples{
check_normality(data,"classif",c("experiment","line"),"feature1","TRUE", output="feature1_qqplot")
}
